Demonstrate: Ceylon type system is Chomsky-2-complete
This removes the previous commit’s restriction of having to explicitly
specify the type arguments. The compiler wasn’t able to infer the
RestStak type argument, so we give the function a regular argument of
that type.

Using this technique, it should also be possible to construct a pushdown
automaton with two stacks – that is, a Turing machine. This means that
the Ceylon type system is (most likely) TURING COMPLETE (with the
restriction that you can’t make the typechecker loop, so the number of
steps is hard-coded by the number of function invocations you make).