README: Add SHAs to file links
I might update the examples later (for example, ditching my own B and P box types in favor of regular tuples), and it would be bad if when I didn’t update the README, the links would point to source code that is no longer consistent with it. Now the links always point to the versions of the files at the time that part of the README was written, and if I update the README, then I can update these links as well.